feat: add comments

+++ apps/src-tauri/src/main.rs
@@ -12,7 +12,7 @@ mod daemon;
 
 #[tokio::main]
 async fn main() {
-    // check if debug mode
+    // check if the app is running in debug mode
     if cfg!(debug_assertions) {
         std::env::set_var("RUST_LOG", "debug");
     }
@@ -25,29 +25,35 @@ async fn main() {
     let db_path = tcore::utils::dirs::get_twidge_dir()
         .unwrap()
         .join("library.db");
-    let glob_app = Arc::new(Mutex::new(None));
 
-    std::fs::File::create(db_path.clone()).unwrap();
+    // check if ${twidge_dir}/library.db exists if not create it
+    if !db_path.exists() {
+        std::fs::File::create(db_path.clone()).unwrap();
+    }
 
+    // instantiate a new prisma client and Arc it so that rust doesn't scold me
     let prisma = Arc::new(
         prisma::new_client_with_url(&format!("file:{}", db_path.display()))
             .await
             .unwrap(),
     );
 
-    let glob_app_cloned = glob_app.clone();
-    
+    // used to store the tauri::Window object for the daemon to communicate
+    let global_app = Arc::new(Mutex::new(None));
+    let global_app_cloned = global_app.clone();
+
+    // spawn the daemon in a separate thread
     tokio::spawn(async move {
-        daemon::setup_daemon(glob_app_cloned).await;
+        daemon::setup_daemon(global_app_cloned).await;
     });
 
     tauri::Builder::default()
         .setup(move |app| {
             let window = app.get_window("main").unwrap();
-            set_shadow(&window, true).expect("Failed to set shadow");
+            set_shadow(&window, true);
 
             let window = window.clone();
-            glob_app.lock().unwrap().replace(window);
+            global_app.lock().unwrap().replace(window);
             Ok(())
         })
         .plugin(rspc::integrations::tauri::plugin(router, move || Context {
